feat: bunch of inits
