+++ src/types/config.ts
@@ -41,6 +41,7 @@ export const UnassignConfigSchema = Type.Object({
 export const ModeSchema = Type.Object({
   autoPayMode: Type.Boolean(),
   analyticsMode: Type.Boolean(),
+  incentiveMode: Type.Boolean(),
 });
 
 export const AssignSchema = Type.Object({
